docs-config.ts & moved contactus.ts api file
